refactor: use tab info rather than documentUrl
Until this commit, this plugin uses the property `documentUrl`
of the details passed to webRequest.onBeforeRequest listener.

Yet, this option works only in Firefox 57 or higher.
It is not even documented on MDN:
https://developer.mozilla.org/en-US/Add-ons/WebExtensions/API/webRequest/onBeforeRequest

To make this extension compatible with Chrome, `documentUrl` can't be trusted.

This commit makes the extension not rely on `documentUrl` anymore,
making it use tab information instead.
Rather than questioning the `documentUrl` of the webRequest,
this analyzes the url of the tab making the request.

This commit alos includes:

* minor refactors
* tests for the refactored parts